+++ src/gui/QtThermalFIST/CosmicEoStab.cpp
@@ -1307,8 +1307,10 @@ void CosmicEoSTab::showEoSTable() {
 void CosmicEoSTab::recomputeCalcTable() {
   calcTable.clear();
 
+  int tsize = fCurrentSize;
+
   calcTable.parameter_name = getParameterName();
-  for(int i = 0; i < varvalues.size(); ++i) {
+  for(int i = 0; i < tsize; ++i) {
     calcTable.parameter_values.push_back(varvalues[i]);
     calcTable.temperature_values.push_back(paramsTD[i].T);
   }
@@ -1328,7 +1330,7 @@ void CosmicEoSTab::recomputeCalcTable() {
     calcTable.densities_names.push_back(QString::fromStdString(model->TPS()->Particles()[ic].Name()));
   }
 
-  for(int ir = 0; ir < varvalues.size(); ++ir) {
+  for(int ir = 0; ir < tsize; ++ir) {
     calcTable.densities_values.push_back(paramsTDHRG[ir].densities);
   }
 
@@ -1336,7 +1338,7 @@ void CosmicEoSTab::recomputeCalcTable() {
     calcTable.quantities_names.push_back(QString::fromStdString(
             "n(" + cosmos->GetSpeciesName(ilep) + ")/T³"));
     std::vector<double> lepton_densities;
-    for(int ival = 0; ival < varvalues.size(); ++ival) {
+    for(int ival = 0; ival < tsize; ++ival) {
       lepton_densities.push_back(paramsTD[ival].densities[ilep] / pow(paramsTD[ival].T * xMath::GeVtoifm(), 3.));
     }
     calcTable.quantities_values.push_back(lepton_densities);