+++ src/library/HRGEV/ThermalModelEVDiagonal.cpp
@@ -278,6 +278,7 @@ namespace thermalfist {
       m_wprim[i] = m_PrimCorrel[i][i];
       if (m_densities[i] > 0.) m_wprim[i] *= m_Parameters.T / m_densities[i];
       else m_wprim[i] = 1.;
+      if (m_wprim[i] != m_wprim[i]) m_wprim[i] = 1.;
     }
 
     m_TwoParticleCorrelationsCalculated = true;